requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-jazzmin==3.0.1
django-storages==1.14.4
jmespath==1.0.1
mysql-connector-python==9.1.0
mysqlclient==2.2.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.9.0
